V1.1 - The touch
In the current version, the home page has undergone a modification where all physical buttons and limit switches are initially disabled. The initiation of the game now exclusively relies on clicking the on-screen button. Once activated, all components become enabled, allowing the game to commence.

A notable addition is the introduction of a 'touch sensor' designed for resetting the game state. This sensor facilitates the following actions upon activation:

- The cursor resets to the beginning.
- The question counter resets to zero.
- The website reverts to the home page.

This enhancement streamlines the user experience and provides a convenient mechanism for resetting the game's parameters.