V1.2 - The front refactor
In this latest update, significant enhancements have been made to the website's pages, encompassing improvements in colors, fonts, dimensions, and overall aesthetics.

A notable addition is the implementation of an end screen that dynamically adjusts based on the player's outcome. A cookie now stores a boolean value indicating whether the player succeeded or failed in the game. Success is determined by the player successfully activating the right limit switch, while failure occurs if the player answers all questions incorrectly and exhausts their question stock.

On the end screen, only the red button is active, preventing unintended interactions. The right limit switch remains enabled on this screen to prevent a potential bug where, if the player completes their last question (resulting in failure) and simultaneously activates the right limit switch, it could cause unexpected behavior.

A refined feature has been introduced to address cursor interactions on the end screen. Instead of immediately returning to the beginning, the cursor now briefly moves backward for 0.5 seconds before waiting for the player to click the red button to return to the home screen.

Furthermore, a new addition to the game is a chronometer that tracks real-time values, stored in a separate cookie. This chronometer serves the purpose of displaying the elapsed time on the end screen.

To address issues with the checkTimeout function, it has been removed and replaced with a physical switch mechanism to disable motors, providing a more robust solution.